requirements.txt
Django==1.11.9

django-braces==1.11.0
django-celery==3.2.1
django-compressor==2.1
django-extensions==1.7.9
celery==3.1.24
flower==0.9.1
gunicorn==19.7.1
psycopg2==2.7.3.2
youtube-dl>=2014.06.26
